ec2-fabfile.py

+++ deploy/ec2-fabfile.py
@@ -55,7 +55,7 @@ if 'xxx' == os.environ.get('AWS_SECRET').lower():
 ec2 = boto.ec2.connect_to_region('us-east-1', 
         aws_access_key_id = os.environ.get('AWS_ID'),
         aws_secret_access_key = os.environ.get('AWS_SECRET'))
-env.use_ssh_jonfig = True
+env.use_ssh_config = True
 env.disable_known_hosts = True
 env.connection_attempts = True
 env.timeout = 40
@@ -78,8 +78,6 @@ def create_box():
     print(new_instance.public_dns_name)
     return new_instance
 
-def old_box(public_dns_name):
-    return [i for i in ec2.get_only_instances() if i.public_dns_name == public_dns_name][0]
 
 def create_keypair(source = AMI_ID+'-amfarrell'):
     try:
@@ -88,7 +86,7 @@ def create_keypair(source = AMI_ID+'-amfarrell'):
         pass
 
     kp = ec2.create_key_pair(source)
-    filename = os.environ.get('EC2_KEY_PATH', './ec2-{}.key'.format(datetime.datetime.now().strftime('%Y-%m-%d_%H:%M')))
+    filename = os.environ.get('EC2_KEY_PATH', './ec2-{}.pem'.format(datetime.datetime.now().strftime('%Y-%m-%d_%H:%M')))
     kfile = open(filename, 'wb')
     def file_mode(user, group, other):
         return user*(8**2) + group*(8**1) + other*(8**0)
@@ -115,8 +113,20 @@ def test_ssh(instance, key_file):
     log.debug('Trying to connect...')
     run('pwd')
 
+def connect_to_existing_machine(ip, key_file_path):
+    env.user = 'ubuntu'
+    env.hosts = ['{}@{}'.format(env.user, ip)]
+    env.host = '{}@{}'.format(env.user, ip)
+    env.host_string = '{}@{}'.format(env.user, ip)
+    env.key_file = key_file_path
+    env.key_filename = key_file_path
+    env.forward_agent = True
+    log.info('Key file: %s' % (key_file_path))
+    log.debug('Trying to connect...')
+    run('pwd')
 
-def apt_installs(instance):
+
+def apt_installs():
     log.info("installing packages with apt-get")
     sudo("add-apt-repository -y ppa:keithw/mosh")
     sudo("apt-get update -y")
@@ -130,13 +140,13 @@ def apt_installs(instance):
         'tig']
     sudo("apt-get install -y {}".format(' '.join(packages)))
 
-def fix_sshd_config(instance):
+def fix_sshd_config():
     '''root needs an actual shell, so fix the sshd_config.'''
     config_file = '/etc/ssh/sshd_config'
     uncomment(config_file, r'^.*PermitRootLogin yes', use_sudo=True)
     comment(config_file, r'^PermitRootLogin forced-commands-only', use_sudo=True)
 
-def install_miniconda(instance):
+def install_miniconda():
     url = 'http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh'
     run("wget {}".format(url))
     run("chmod +x ./Miniconda-latest-Linux-x86_64.sh")
@@ -144,67 +154,74 @@ def install_miniconda(instance):
     run("echo 'export PATH=/home/ubuntu/miniconda/bin:\$PATH' >> ~/.bashrc")
     run("source ~/.bashrc")
 
-def install_repo(instance):
+def install_repo(public_dns_name, ip_address):
     url = 'https://github.com/memex-explorer/memex-explorer/'
     if os.environ.get('GIT_BRANCH'):
         run("git clone {} --branch {}".format(url, os.environ.get('GIT_BRANCH')))
     else:
         run("git clone {}".format(url))
     run("~/miniconda/bin/conda env update --name root --file ~/memex-explorer/environment.yml")
-    run("echo 'HOSTNAME = \"{}\"' >> {}".format(instance.public_dns_name, SETTINGS_FILENAME))
+    run("echo 'HOSTNAME = \"{}\"' >> {}".format(public_dns_name, SETTINGS_FILENAME))
     run("echo 'ROOT_PORT = \"{}\"' >> {}".format(MEMEX_APP_PORT, SETTINGS_FILENAME))
-    run("echo 'IP_ADDR = \"{}\"' >> {}".format(instance.ip_address, SETTINGS_FILENAME))
+    run("echo 'IP_ADDR = \"{}\"' >> {}".format(ip_address, SETTINGS_FILENAME))
     run("~/miniconda/bin/python ~/memex-explorer/source/manage.py migrate")
     run("~/miniconda/bin/python ~/memex-explorer/source/manage.py create_apps_Tika_ES_Kibana")
     sudo("ln -s ~/miniconda/bin/docker-compose /bin/docker-compose")
 
-def start_nginx(instance):
+def start_nginx():
     sudo("~/miniconda/bin/python ~/memex-explorer/source/manage.py refresh_nginx")
 
-def install_docker(instance):
+def install_docker():
     run("chmod +x ~/memex-explorer/deploy/install-docker.sh")
     run("~/memex-explorer/deploy/install-docker.sh")
     sudo("docker pull elasticsearch")
     sudo("docker pull continuumio/tika")
     sudo("docker pull continuumio/kibana")
 
-def conventience_aliases(instance):
+def conventience_aliases():
     run("echo 'alias dj=\"~/miniconda/bin/python ~/memex-explorer/source/manage.py\"' >> ~/.bashrc")
 
-def start_server_running(instance):
-    run("redis-server &")
-    run("disown")
-    run("celery --workdir=\"$HOME/memex-explorer/source\" -A memex worker &")
-    run("disown")
+def start_server_running():
+    run("redis-server & disown")
+    run("celery --workdir=\"$HOME/memex-explorer/source\" -A memex worker & disown")
     run("~/miniconda/bin/python ~/memex-explorer/source/manage.py runserver 0.0.0.0:{}".format(MEMEX_APP_PORT))
 
 
 
-key_filename = create_keypair()
-instance = create_box()
-subprocess.check_output(['cp', key_filename, 
-                         os.path.join(os.path.dirname(key_filename), 'ec2-{}.key'.format(instance.ip_address))])
-ssh_command = 'ssh -i {key} ubuntu@{ip} "'.format(ip=instance.ip_address, key=key_filename)
-mosh_command = 'mosh ubuntu@{ip} --ssh="ssh -i {key}"'.format(ip=instance.ip_address, key=key_filename)
+if os.environ.get('MEMEX_IP_ADDR'):
+    ip_address = os.environ.get('MEMEX_IP_ADDR')
+    key_filename = os.path.abspath('./ec2-{}.pem'.format(ip_address))
+    public_dns_name = 'ec2-{}.compute-1.amazonaws.com'.format(ip_address.replace('.','-'))
+    connect_to_existing_machine(ip_address, key_filename)
+else:
+    key_filename = create_keypair()
+    instance = create_box()
+    subprocess.check_output(['cp', key_filename, 
+                             os.path.join(os.path.dirname(key_filename), 'ec2-{}.pem'.format(instance.ip_address))])
+    public_dns_name = instance.public_dns_name
+    ip_address = instance.ip_address
+    test_ssh(instance, key_filename)
+ssh_command = 'ssh -i {key} ubuntu@{ip} "'.format(ip=ip_address, key=key_filename)
+mosh_command = 'mosh ubuntu@{ip} --ssh="ssh -i {key}"'.format(ip=ip_address, key=key_filename)
 if 'quitafterec2spinup' in sys.argv:
     print(ssh_command)
     quit()
 try:
-    test_ssh(instance, key_filename)
     print(ssh_command)
-    apt_installs(instance)
+    apt_installs()
     print(mosh_command)
-    fix_sshd_config(instance)
+    fix_sshd_config()
 except Exception:
-    print("{} failed!".format(instance.public_dns_name))
-    ec2.terminate_instances([instance.id])
+    print("{} failed!".format(public_dns_name))
+    if not os.environ.get('MEMEX_IP_ADDR'):
+        ec2.terminate_instances([instance.id])
     raise
 try:
-    install_miniconda(instance)
-    install_repo(instance)
-    start_nginx(instance)
-    install_docker(instance)
-    start_server_running(instance)
+    install_miniconda()
+    install_repo()
+    start_nginx()
+    install_docker()
+    start_server_running()
 except Exception:
     print(ssh_command)
     print(mosh_command)
